+++ R/posixt.R
@@ -967,10 +967,10 @@ date_month_factor.POSIXt <- function(x,
 #'
 #' `date_format()` formats a date-time (POSIXct) using a `format` string.
 #'
-#' If `format` is `NULL`, a default format of `"%Y-%m-%d %H:%M:%S%Ez[%Z]"` is
-#' used. This maximizes the chance for constructing a string that can be
-#' reproducibly parsed into a valid date-time using
-#' [date_time_parse_complete()].
+#' If `format` is `NULL`, a default format of `"%Y-%m-%dT%H:%M:%S%Ez[%Z]"` is
+#' used. This matches the default format that [date_time_parse_complete()]
+#' parses. Additionally, this format matches the de-facto standard extension to
+#' RFC 3339 for creating completely unambiguous date-times.
 #'
 #' @inheritParams ellipsis::dots_empty
 #' @inheritParams format.clock_zoned_time
@@ -992,6 +992,9 @@ date_month_factor.POSIXt <- function(x,
 #'
 #' # Default
 #' date_format(x)
+#'
+#' # Which is parseable by `date_time_parse_complete()`
+#' date_time_parse_complete(date_format(x))
 #'
 #' date_format(x, format = "%B %d, %Y %H:%M:%S")
 #'
@@ -1134,14 +1137,17 @@ date_set_zone.POSIXt <- function(x, zone) {
 #' ## date_time_parse_complete()
 #'
 #' `date_time_parse_complete()` is a parser for _complete_ date-time strings,
-#' like `"2019-01-01 00:00:00-05:00[America/New_York]"`. A complete date-time
+#' like `"2019-01-01T00:00:00-05:00[America/New_York]"`. A complete date-time
 #' string has both the time zone offset and full time zone name in the string,
 #' which is the only way for the string itself to contain all of the information
-#' required to construct a zoned-time. Because of this,
+#' required to unambiguously construct a zoned-time. Because of this,
 #' `date_time_parse_complete()` requires both the `%z` and `%Z` commands to be
 #' supplied in the `format` string.
 #'
-#' The default `format` used is `"%Y-%m-%d %H:%M:%S%Ez[%Z]"`.
+#' The default `format` used is `"%Y-%m-%dT%H:%M:%S%Ez[%Z]"`. This matches the
+#' default result from calling `date_format()` on a POSIXct date-time.
+#' Additionally, this format matches the de-facto standard extension to RFC 3339
+#' for creating completely unambiguous date-times.
 #'
 #' ## date_time_parse_abbrev()
 #'
@@ -1192,7 +1198,7 @@ date_set_zone.POSIXt <- function(x, zone) {
 #' # Same time as above, except this is a completely unambiguous parse that
 #' # doesn't require a `zone` argument, because the zone name and offset are
 #' # both present in the string
-#' date_time_parse_complete("2020-01-01 05:06:07-05:00[America/New_York]")
+#' date_time_parse_complete("2020-01-01T05:06:07-05:00[America/New_York]")
 #'
 #' # Only day components
 #' date_time_parse("2020-01-01", "America/New_York", format = "%Y-%m-%d")
@@ -1219,8 +1225,8 @@ date_set_zone.POSIXt <- function(x, zone) {
 #' # that the offset and zone name are both in the string, which resolves
 #' # the ambiguity
 #' complete_times <- c(
-#'   "1970-10-25 01:00:00-04:00[America/New_York]",
-#'   "1970-10-25 01:00:00-05:00[America/New_York]"
+#'   "1970-10-25T01:00:00-04:00[America/New_York]",
+#'   "1970-10-25T01:00:00-05:00[America/New_York]"
 #' )
 #' date_time_parse_complete(complete_times)
 #'
@@ -1239,7 +1245,7 @@ date_set_zone.POSIXt <- function(x, zone) {
 #' # seconds, first parse them into a clock type that can handle sub-seconds to
 #' # fully capture that information, then round using whatever convention is
 #' # required for your use case before converting to a date-time.
-#' x <- c("2019-01-01 00:00:01.1", "2019-01-01 00:00:01.78")
+#' x <- c("2019-01-01T00:00:01.1", "2019-01-01T00:00:01.78")
 #'
 #' x <- naive_time_parse(x, precision = "millisecond")
 #' x