+++ NEWS.md
@@ -4,6 +4,13 @@
   in many of the date formatting and parsing functions have been slightly
   altered. The following changes have been made:
   
+  * Date-times (POSIXct):
+  
+    * `date_format()` now print a `T` between the date and time.
+    
+    * `date_time_parse_complete()` now expects a `T` between the date and time
+      by default.
+  
   * Sys-times:
   
     * `format()` and `as.character()` now print a `T` between the date and time.
@@ -16,6 +23,13 @@
     
     * `naive_time_parse()` now expects a `T` between the date and time by
       default.
+      
+  * Zoned-times:
+  
+    * `format()` and `as.character()` now print a `T` between the date and time.
+    
+    * `zoned_time_parse_complete()` now expects a `T` between the date and time
+      by default.
 
 * Further improved documentation of undefined behavior resulting from attempting
   to parse sub-daily components of a string that is intended to be parsed into